Split DrawColourInfo into its own class
